temporary debug code: add implementation of alternative approaches
