+++ plugins/org.eclipse.n4js.xtext.ide/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.emf.ecore.change,
  org.eclipse.lsp4j,
  org.eclipse.lsp4j.jsonrpc,
- org.eclipse.n4js.smith,
  org.eclipse.n4js.utils,
  org.apache.log4j,
  com.google.inject,
